webapi/webapi-presentation-w3c-tests/presentation/Presentation_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>Presentation basic attributes</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Chunyan Wang" href="mailto:chunyanx.wang@intel.com">
<link rel="help" href="https://w3c.github.io/presentation-api/#interface-presentation">
<link rel="help" href="https://w3c.github.io/presentation-api/#idl-def-presentationrequest">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

test(function() {
  assert_true("presentation" in navigator);
  assert_true("defaultRequest" in navigator.presentation);
}, "Check that the Presentation API is present");

test(function() {
  assert_equals(typeof(navigator.presentation), "object");
  assert_equals(typeof(navigator.presentation.defaultRequest), "object");
}, "Check that the Presentation API property types.");

test(function() {
  assert_true("PresentationRequest" in window);
  var request = new PresentationRequest("display/presentation.html");
  assert_true("start" in request);
  assert_true("reconnect" in request);
  assert_true("getAvailability" in request);
  assert_true("onconnectionavailable" in request);
}, "Check that the PresentationRequest API is present");

test(function() {
  var request = new PresentationRequest("display/presentation.html");
  assert_equals(typeof(request.start), "function");
  assert_equals(typeof(request.reconnect), "function");
  assert_equals(typeof(request.getAvailability), "function");
  assert_equals(typeof(request.onconnectionavailable), "object");
}, "Check that the PresentationRequest API property types.");

</script>
